# backend/templates/openapi-schema.yml
openapi: 3.0.2
info:
  title: "RMС REST API"
  version: "1.0.0"
  description: ""
paths:
  /api/users/:
    get:
      operationId: Список пользователей
      description: ""
      security:
        - Token: [ ]
      parameters:
        - name: limit
          required: false
          in: query
          description: "Количество объектов на странице"
          schema:
            type: integer
        - name: offset
          required: false
          in: query
          description: "Смещение объектов по выдаче"
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                    description: "Общее количество объектов в базе"
                  next:
                    type: string
                    nullable: true
                    format: uri
                    example: http://rest-api.krasrm.com/api/users/?limit=1&offset=3
                    description: "Ссылка на следующую страницу"
                  previous:
                    type: string
                    nullable: true
                    format: uri
                    example: http://rest-api.krasrm.com/api/users/?limit=1&offset=1
                    description: "Ссылка на предыдущую страницу"
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                    description: "Список объектов текущей страницы"
          description: "OK"
        '401':
          $ref: "#/components/responses/AuthenticationError"
        '403':
          $ref: "#/components/responses/PermissionDenied"
      tags:
        - Пользователи
    post:
      operationId: Регистрация пользователя
      description: ""
      security:
        - Token: [ ]
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomUserCreate"
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: "Пользователь успешно создан"
        '400':
          $ref: "#/components/responses/ValidationError"
      tags:
        - Пользователи
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          example: 2
        email:
          type: string
          nullable: true
          format: email
          maxLength: 254
          description: "Адрес электронной почты"
          example: "ivanovi@krasrm.com"
        username:
          type: string
          description: "Уникальный юзернейм"
          pattern: ^[\w.@+-]+\z
          maxLength: 150
          example: "ivanovi"
        password:
          type: string
          description: "Пароль пользователя в SHA256"
          example: "12345"
        last_name:
          type: string
          maxLength: 150
          description: "Фамилия"
          example: "Иванов"
        first_name:
          type: string
          maxLength: 150
          description: "Имя"
          example: "Иван"
        middle_name:
          type: string
          nullable: true
          maxLength: 150
          description: "Отчество"
          example: "Иванович"
        role:
          type: string
          description: "Роль"
          example: "Менеджер"
        phone_number:
          type: string
          nullable: true
          format: phone number
          description: "Номер телефона"
          example: "8(800) 222-59-38"
        is_active:
          type: bool
          description: "Актуальность пользователя"
          example: True
        created:
          type: timestamp
          description: "Дата регистрации"
          example: "01-01-2001 16:20:00"
      required:
        - username
        - password
        - last_name
        - first_name
        - role
        - is_active
...
    CustomUserCreate:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
          description: "Email пользователя"
          example: "pupkinv@krasrm.com"
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Уникальный юзернейм"
          pattern: ^[\w.@+-]+\z
          maxLength: 150
          example: "vasya.pupkin"
        first_name:
          type: string
          maxLength: 150
          description: "Имя"
          example: "Вася"
        last_name:
          type: string
          maxLength: 150
          description: "Фамилия"
          example: "Пупкин"
        password:
          type: string
          maxLength: 150
          description: "Пароль"
          example: "Qwerty123"
      required:
        - username
        - password
        - first_name
        - last_name
        - email
...
    ValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        field_name:
          description: 'Название поля, в котором произошли ошибки. Таких полей может быть несколько'
          example: [ 'Обязательное поле.' ]
          type: array
          items:
            type: string
...
    AuthenticationError:
      description: 'Пользователь не авторизован'
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
    PermissionDenied:
      description: 'Недостаточно прав'
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "У вас недостаточно прав для выполнения данного действия."
          type: string
...
  responses:
    ValidationError:
      description: 'Ошибки валидации в стандартном формате DRF'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    AuthenticationError:
      description: 'Пользователь не авторизован'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
    PermissionDenied:
      description: 'Недостаточно прав'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PermissionDenied'
...
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token
...